\name{optionPriceR-package}
\alias{optionPriceR-package}
\alias{optionPriceR}
\docType{package}
\title{
  A package for pricing of an European style down-and-out call option
}
\description{
  The package calculates price paths using Monte Carlo Simulation and then it sets the price to an average of all the paths that ended above the strike and were above the barrier.
}
\details{
  Only one function is available getDownAndOutCallPrice
}
\author{
Mateusz Buczynski

Maintainer: Mateusz Buczynski <mateusz.buczynski@interia.pl>
}
\keyword{ package }

