\name{optionPricer-package}
\alias{optionPricer-package}
\alias{optionPricer}
\docType{package}
\title{
  Pricing a European Up-and-In Put Option with Monte Carlo similation
}
\description{
  This package simulates a European Up-and-In Put Option and returns its theoretical price.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Dustin Pacholleck
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
